Implemented our own abstraction of Channels, instead of putting that up in the "slack" package.
Added "SendPrivateMessage" which ensures we have an IM channel open with
a user.

Now keeps track of open DM channels, was as with Groups and Channels.

Added "Matches" to the Listener, now does a regexp match on the "msg.Text"
and assigns it to the *Message.Match field.

Fixed a bug with "Message.IsPrivate()" which disregarded the fact that private
messages are actually in a channel, but a "D*" channel.

Implemented ListenReaction on both a Message and a Reply. This follows
the same model as the "Listener".